Cria um creator para escolher o DataSourceFactory a ser utilizado de acordo com as configurações. Issue #31
