+++ src/main/appengine/app.yaml
@@ -5,5 +5,5 @@ manual_scaling:
 
 env_variables:
   JAVA_TOOL_OPTIONS: "-XX:MaxRAM=256m -XX:ActiveProcessorCount=2 -Xmx32m"
-  APPENGINE_ENVIRONMENT: "true"
+  APP_ENGINE_ENVIRONMENT: "true"
   GOOGLE_CLIENT_ID: "gs://register-272219.appspot.com/authentication/client-id-registrant.json"
