+++ src/main/java/br/com/breadware/configuration/BeanNames.java
@@ -8,10 +8,6 @@ public final class BeanNames {
 
     public static final String JSON_FACTORY = "jsonFactory";
 
-    public static final String HTTP_REQUEST_INITIALIZER = "httpRequestInitializer";
-
-    public static final String HTTP_REQUEST_INITIALIZER_CREATOR = "httpRequestInitializerCreator";
-
     public static final String MESSAGE_SOURCE = "messageSource";
 
     public static final String MESSAGE_RETRIEVER = "messageRetriever";
@@ -37,6 +33,8 @@ public final class BeanNames {
     public static final String STORAGE = "storage";
 
     public static final String DATA_STORE_FACTORY = "cloudStorageDataFactory";
+
+    public static final String CREDENTIAL = "credential";
 
     private BeanNames() {
         // Private constructor to avoid object instantiation.
