+++ src/main/java/br/com/breadware/configuration/RegistrantConfiguration.java
@@ -1,19 +1,15 @@
 package br.com.breadware.configuration;
 
-import br.com.breadware.bo.CloudStorageBo;
-import br.com.breadware.google.cloud.httprequestinitializer.datastore.CloudStorageDataStoreFactory;
-import br.com.breadware.model.CloudStorageLocation;
+import br.com.breadware.google.cloud.authorization.datastore.DataStoreFactoryCreator;
 import br.com.breadware.properties.RegistrantProperties;
 import br.com.breadware.properties.google.GcpAuthorizationProperties;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.api.client.util.store.DataStoreFactory;
-import com.google.api.client.util.store.FileDataStoreFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import java.io.File;
 import java.io.IOException;
 import java.time.ZoneId;
 import java.util.Locale;
@@ -42,17 +38,7 @@ public class RegistrantConfiguration {
     }
 
     @Bean(BeanNames.DATA_STORE_FACTORY)
-    public DataStoreFactory createDataStoreFactory(CloudStorageBo cloudStorageBo, GcpAuthorizationProperties gcpAuthorizationProperties) throws IOException {
-        DataStoreFactory dataStoreFactory;
-
-        String tokensStorageLocation = gcpAuthorizationProperties.getTokensStoragePath();
-        if (cloudStorageBo.isValidObjectUri(tokensStorageLocation)) {
-            CloudStorageLocation clientIdCloudStorageLocation = cloudStorageBo.mapAsCloudStorageLocation(tokensStorageLocation);
-            dataStoreFactory = new CloudStorageDataStoreFactory(cloudStorageBo, clientIdCloudStorageLocation);
-        } else {
-            dataStoreFactory = new FileDataStoreFactory(new File(tokensStorageLocation));
-        }
-        LOGGER.info("Created a \"{}\" object to attend \"{}\" inject beans.", dataStoreFactory.getClass(), DataStoreFactory.class);
-        return dataStoreFactory;
+    public DataStoreFactory createDataStoreFactory(DataStoreFactoryCreator dataStoreFactoryCreator, GcpAuthorizationProperties gcpAuthorizationProperties) throws IOException {
+        return dataStoreFactoryCreator.create(gcpAuthorizationProperties.getTokensStoragePath());
     }
 }
