--- src/main/java/br/com/breadware/google/cloud/httprequestinitializer/HttpRequestInitializerCreator.java
@@ -1,8 +1,0 @@
-package br.com.breadware.google.cloud.httprequestinitializer;
-
-import com.google.api.client.http.HttpRequestInitializer;
-
-public interface HttpRequestInitializerCreator {
-
-  HttpRequestInitializer create();
-}
