+++ src/main/java/br/com/breadware/model/message/ErrorMessage.java
@@ -32,7 +32,9 @@ public enum ErrorMessage implements Message {
     ERROR_WHILE_HANDLING_GMAIL_MESSAGE(26),
     TEXT_IS_NOT_CLOUD_STORAGE_URI(27),
     ERROR_RETRIEVING_GOOGLE_CLIENT_ID(28),
-    COULD_NOT_LOCATE_CLIENT_ID_FILE(29);
+    COULD_NOT_LOCATE_CLIENT_ID_FILE(29),
+    OBJECT_ON_CLOUD_STORAGE_IS_NOT_HASHMAP(30),
+    ERROR_RETRIEVING_DATA_MAP_FROM_CLOUD_STORAGE(31);
 
     public static final String FILE_PATH = "error/error".replace("/", File.separator);
