+++ src/main/java/br/com/breadware/model/message/LoggerMessage.java
@@ -14,8 +14,7 @@ public enum LoggerMessage implements Message {
     MESSAGE_BODY_IS_INVALID_JSON(8),
     MESSAGE_IS_NOT_ASSOCIATE_INFORMATION(9),
     GMAIL_MESSAGE_HAS_ALREADY_BEEN_HANDLED(10),
-    GMAIL_MESSAGE_HAS_BEEN_HANDLED(11),
-    SKIPPING_ENVIRONMENT_VARIABLE_CHECK_ON_APP_ENGINE(12);
+    GMAIL_MESSAGE_HAS_BEEN_HANDLED(11);
 
     public static final String FILE_PATH = "logger/logger".replace("/", File.separator);
