+++ src/main/java/br/com/breadware/util/EnvironmentVariables.java
@@ -1,11 +1,13 @@
 package br.com.breadware.util;
 
-public class EnvironmentVariables {
+public final class EnvironmentVariables {
 
-  public static final String APPENGINE_ENVIRONMENT = "APPENGINE_ENVIRONMENT";
-
-  public static final String GOOGLE_CLOUD_PROJECT = "GOOGLE_CLOUD_PROJECT";
+  public static final String APP_ENGINE_ENVIRONMENT = "APP_ENGINE_ENVIRONMENT";
 
   public static final String CLIENT_ID_FILE_LOCATION = "GOOGLE_CLIENT_ID";
+
+  private EnvironmentVariables() {
+    // Private constructor to avoid instantiation.
+  }
 
 }
