+++ src/main/resources/config/application-develop.yml
@@ -6,4 +6,4 @@ registrant:
   google:
     cloud-platform:
       authorization:
-        tokens-storage-path: "tokens"
+        tokens-storage-path: "tokens/"
