+++ src/main/resources/i18n/messages/error/error.properties
@@ -27,3 +27,5 @@ error.26=Error while handling GMail message "{0}".
 error.27=Text "{0}" is not a valid Google Cloud Storage object URI.
 error.28=Error retrieving Google Client ID file.
 error.29=Could not locate Google Client ID file "{0}".
+error.30=Object retrieved from Cloud Storage is not a HashMap.
+error.31=Error retrieving data map from Google Cloud Storage.
