+++ src/main/resources/i18n/messages/error/error_pt_BR.properties
@@ -27,3 +27,5 @@ error.26=Erro ao analisar a mensagem "{0}" do GMail.
 error.27=O texto "{0}" não representa a URI de um objeto armazenado no Google Cloud Storage.
 error.28=Erro ao obter o arquivo de identificação do cliente no Google.
 error.29=Não foi possível localizar o arquivo de identificação do cliente no Google "{0}".
+error.30=Objeto armazenado no Cloud Storage não é um HashMap.
+error.31=Erro ao obter o mapa de dados do Google Cloud Storage.
