+++ src/main/resources/i18n/messages/logger/logger.properties
@@ -9,4 +9,3 @@ logger.8=E-mail body content is not a valid Json. Partial message content: "{0}"
 logger.9=GMail message {0} does not appear to have associate information. Skipping its content.
 logger.10=GMail message {0} has already been handled.
 logger.11=GMail message {0} hand been successfully handled.
-logger.12=Program is running on Google App Engine environment. Skipping {0} environment variable check.
