+++ src/main/resources/i18n/messages/logger/logger_pt_BR.properties
@@ -9,4 +9,3 @@ logger.8=O conteúdo do corpo do e-mail não é um Json válido. Conteúdo parci
 logger.9=A mensagem GMail {0} não parece ser a respeito de informações de um associado. Ignorando o seu conteúdo.
 logger.10=A mensagem GMail de ID {0} já foi analisada.
 logger.11=A mensagem GMail {0} foi analisada com sucesso.
-logger.12=O programa está sendo executado no ambiente do Google App Engine. Ignorando a verificação da variável de ambiente {0}.
